requirements.txt
bcrypt==3.2.0
beautifulsoup4==4.10.0
black==21.9b0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
coverage==6.0.2
flash==1.0.3
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Ext==0.1
Flask-Login==0.3.2
Flask-SQLAlchemy==2.1
Flask-WTF==0.15.1
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
lyricsgenius==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
psycopg2-binary==2.9.1
pycparser==2.20
python-dotenv==0.19.0
regex==2021.10.8
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
spotipy==2.19.0
SQLAlchemy==1.3.23
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
WTForms==2.3.3
